+++ benchmarks/benches/crypto_bench.rs
@@ -194,7 +194,7 @@ pub fn room_key_sharing(c: &mut Criterion) {
                 machine.mark_request_as_sent(&request.txn_id, &to_device_response).await.unwrap();
             }
 
-            machine.invalidate_group_session(room_id).await.unwrap();
+            machine.discard_room_key(room_id).await.unwrap();
         })
     });
 
@@ -225,7 +225,7 @@ pub fn room_key_sharing(c: &mut Criterion) {
                 machine.mark_request_as_sent(&request.txn_id, &to_device_response).await.unwrap();
             }
 
-            machine.invalidate_group_session(room_id).await.unwrap();
+            machine.discard_room_key(room_id).await.unwrap();
         })
     });